requirements.txt
fastapi
uvicorn
requests
deepgram-sdk
python-multipart